# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeSync",
    "description": "Real-time collaborative coding platform API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Create a new interview session",
        "description": "Creates a new collaborative coding session and returns the session details with a shareable URL.",
        "operationId": "create_session_api_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "Get session details",
        "description": "Retrieve the current state of an interview session by its ID.",
        "operationId": "get_session_api_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sessions"
        ],
        "summary": "Delete a session",
        "description": "Permanently delete an interview session.",
        "operationId": "delete_session_api_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "description": "Health check endpoint for monitoring.\n\nReturns:\n    Status indicating the service is healthy.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SessionCreateRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "default": "python"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": ""
          }
        },
        "type": "object",
        "title": "SessionCreateRequest",
        "description": "Request body for creating a new session.\n\nAttributes:\n    language: Programming language for the session.\n    code: Initial code content."
      },
      "SessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "participant_count": {
            "type": "integer",
            "title": "Participant Count"
          },
          "share_url": {
            "type": "string",
            "title": "Share Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "language",
          "participant_count",
          "share_url"
        ],
        "title": "SessionResponse",
        "description": "Response model for session data.\n\nAttributes:\n    id: Unique session identifier.\n    code: Current code content.\n    language: Programming language.\n    participant_count: Number of connected participants.\n    share_url: URL for sharing the session."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}